% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts}
\alias{counts}
\title{Number of deaths}
\format{
A data frame with 950 rows and 5 variables:
\describe{
\item{year}{Year of death}
\item{disease}{Cause of death}
\item{age_group}{5-year age group}
\item{sex}{Sex}
\item{counts}{Number of deaths}
}
}
\usage{
counts
}
\description{
A data set containing \emph{simulated} number of deaths by disease, year,
and demographic characteristics.
}
\keyword{datasets}
