% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{death_rates}
\alias{death_rates}
\title{Age-specific death rates}
\format{
A data frame with 95 rows and 3 variables:
\describe{
\item{disease}{Cause of death}
\item{age_group}{5-year age group}
\item{std_rt}{Death rate expressed as number of death per unit of population}
}
}
\usage{
death_rates
}
\description{
A data set containing \emph{simulated} age-specific death rates.
}
\keyword{datasets}
