% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci_old.R
\name{get_ci}
\alias{get_ci}
\title{Calculate a confidence interval}
\usage{
get_ci(dist, interval, estimate, variance = NULL, denominator = NULL)
}
\arguments{
\item{dist}{A string to indicate the type of probability distribution
to follow. Can be one of the following: \code{"normal"}, \code{"log normal"}, or \code{"poisson"}.
This parameter is case insensitive.}

\item{interval}{A scalar, between 0 and 1, indicating the width of the interval.
For example, for a 95\% confidence interval, use \code{interval = 0.95}.}

\item{estimate}{A numeric vector containing the point estimates. When constructing
a confidence interval using the Poisson distribution, \code{estimate} must be a (vector of)
whole number(s). When the Poisson confidence interval is on a \emph{rate}, \code{estimate} is the numerator.}

\item{variance}{A numeric vector containing the variances. This parameter must
be left empty when constructing a confidence interval for a Poisson distribution.}

\item{denominator}{A numeric vector containing the denominator when constructing
a Poisson confidence interval on a \emph{rate}. Default value is
\code{1} when \code{dist = "poisson"}. For other distributions, this parameter must be left empty.}
}
\value{
A data frame with 2 columns \code{upper} and \code{lower} which contain the upper
and lower bounds of a confidence interval respectively.
}
\description{
Calculate a confidence interval given the estimate(s),
the variance(s), the width of the interval, and the type of probability
distribution to follow.
}
\details{
This low-level function doesn't need to know what the point estimates
represent (e.g. rate, ratio, etc.). It also assumes that the estimates and variances
are provided in the same order.

\code{estimate} and \code{variance}, when provided, must have the same length. If \code{denominator} is
provided, its length must be 1 or equal to the length of \code{estimate}. If the length of
\code{denominator} is 1, R's \href{http://www.r-tutor.com/r-introduction/vector/vector-arithmetics}{recycling rules} apply.

The Poisson distribution is implemented using the chi-squared distribution to allow for
the continuous extension of integer variable (\code{estimate}).
}
\examples{
\dontrun{
# calculate a single confidence interval
get_ci(dist = "normal", interval = 0.95, estimate = 10, variance = 4)
get_ci(dist = "log normal", interval = 0.9, estimate = 10, variance = 4)
get_ci(dist = "poisson", interval = 0.99, estimate = 30, denominator = 100)

# calculate multiple confidence intervals
get_ci(dist = "normal", interval = 0.95, estimate = c(10, 20), variance = c(4, 9))

# use a data frame to calculate confidence intervals
estimate <- c(10, 20, 30)
variance <- c(4, 9, 16)
df <- data.frame(estimate, variance)

get_ci(dist = "normal", interval = 0.95, df$estimate, df$variance)
}

}
