% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci_gamma}
\alias{get_ci_gamma}
\title{Calculate a gamma distribution confidence interval}
\usage{
get_ci_gamma(interval, estimate, weights, variance, method = "tcz06")
}
\arguments{
\item{interval}{A scalar between 0 and 1, indicating the width of the interval.}

\item{estimate}{A numeric vector containing the point estimates.}

\item{method}{}
}
\value{
A data frame with columns \code{lower} and \code{upper} which contain the lower
and upper limits of a confidence interval respectively.
}
\description{
This function uses \code{\link[stats:Chisquare]{stats::qchisq()}} to derive approximate confidence
intervals.
}
