% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci_lnorm}
\alias{get_ci_lnorm}
\title{Calculate a log normal distribution confidence interval}
\usage{
get_ci_lnorm(interval, estimate, variance_log)
}
\arguments{
\item{interval}{A scalar between 0 and 1, indicating the width of the
interval. For example, use \code{0.95} to calculate a 95\% confidence interval.}

\item{estimate}{A numeric vector containing the point estimates.}

\item{variance_log}{A numeric vector containing the variances of the
log-transformed \code{estimate}.}
}
\value{
A data frame with columns \code{lower} and \code{upper} which contain the lower
and upper limits of a confidence interval respectively.
}
\description{
Calculate a log normal distribution confidence interval
}
\details{
This function doesn't need to know which type of epidemiological
measure has been passed to \code{estimate} (e.g., rate, risk). It also assumes
that the values of the input arguments \code{estimate} and \code{variance_log} are aligned
and that the vectors are the same length.

This function is used to add confidence intervals in \code{\link[=get_spec_rt]{get_spec_rt()}} and
\code{\link[=get_ds_rt]{get_ds_rt()}}.
}
\examples{
\dontrun{
# calculate a single confidence interval
get_ci_lnorm(interval = 0.95, estimate = 0.2, variance_log = 0.05)

# calculate multiple confidence interval
get_ci_lnorm(interval = 0.95, estimate = c(0.2, 0.4), variance_log = c(0.2, 0.4))

# use columns of a data frame as inputs
df <- data.frame(estimate = c(0.2, 0.4), variance_log = c(0.2, 0.4))

# using dplyr
df \%>\%
  dplyr::mutate(get_ci_lnorm(0.975, estimate, variance_log))

# using base
get_ci_lnorm(interval = 0.975, estimate = df$estimate, variance_log = df$variance_log)
}

}
\references{
\href{http://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_stdrate_details.htm}{The STDRATE Procedure}
}
