% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci_norm}
\alias{get_ci_norm}
\title{Calculate a normal distribution confidence interval}
\usage{
get_ci_norm(interval, estimate, variance)
}
\arguments{
\item{interval}{A scalar between 0 and 1, indicating the width of the
interval. For example, use \code{0.95} to calculate a 95\% confidence interval.}

\item{estimate}{A numeric vector containing the point estimates.}

\item{variance}{A numeric vector containing the variances.}
}
\value{
A data frame with columns \code{lower} and \code{upper} which contain the lower
and upper limits of a confidence interval respectively.
}
\description{
Calculate a normal distribution confidence interval.
}
\details{
This function doesn't need to know which type of epidemiological
measure has been passed to \code{estimate} (e.g., rate, risk). It also assumes
that the values of the input arguments \code{estimate} and \code{variance} are aligned
and that the vectors are the same length.

This function is used to construct confidence intervals in \code{\link[=get_spec_rt]{get_spec_rt()}} and
\code{\link[=get_ds_rt]{get_ds_rt()}}. In general, \emph{the normal distribution should not be used to construct
confidence intervals when the standard deviation is greater than a third of
the mean}. The function throws a warning if this is the case.
}
\examples{
\dontrun{
# calculate a single confidence interval
get_ci_norm(interval = 0.95, estimate = 10, variance = 4)

# calculate confidence intervals around multiple estimates
get_ci_norm(interval = 0.95, estimate = c(10, 20), variance = c(4, 9))

# use columns of a data frame as inputs
df <- data.frame(estimate = c(10, 20, 30), variance = c(4, 9, 16))

# using dplyr
df \%>\%
  dplyr::mutate(get_ci_norm(0.975, estimate, variance))

# using base
get_ci_norm(interval = 0.975, estimate = df$estimate, variance = df$variance)
}

}
\references{
\href{http://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_stdrate_details.htm}{The STDRATE Procedure}
}
\seealso{
\code{\link[=get_ci_lnorm]{get_ci_lnorm()}}, \code{\link[=get_ci_pois]{get_ci_pois()}}, \code{\link[=get_ci_gamma]{get_ci_gamma()}}
}
