% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci_pois}
\alias{get_ci_pois}
\title{Calculate a Poisson distribution confidence interval}
\usage{
get_ci_pois(interval, x, y = NULL)
}
\arguments{
\item{interval}{A scalar between 0 and 1, indicating the width of the interval.
For example, use 0.95 to calculate a 95\% confidence interval.}

\item{y}{}
}
\value{
A data frame with columns \code{lower} and \code{upper} which contain the lower
and upper limits of a confidence interval respectively.
}
\description{
Calculate a Poisson distribution confidence interval. This function
imitates \code{\link[stats:Poisson]{stats::qpois()}} using \code{\link[stats:Chisquare]{stats::qchisq()}} to allow for the continuous
extension of the estimate.
}
\details{
This function is used to calculate confidence intervals in \code{\link[=get_spec_rt]{get_spec_rt()}}.
}
\examples{
\dontrun{
# calculate a single confidence interval
get_ci_pois(interval = 0.95, x = 24)
get_ci_pois(interval = 0.95, x = 49, y = 52)

# use columns of a data frame as inputs
df <- data.frame(x = c(24, 49), y = c(94, 52))

# using dplyr
df \%>\%
  dplyr::mutate(get_ci_pois(0.95, x, y))

# using base
get_ci_pois(0.95, df$x, df$y)
}

}
