% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_dist_name}
\alias{get_dist_name}
\title{Use regex to detect a probability distribution name}
\usage{
get_dist_name(dist)
}
\arguments{
\item{dist}{A string provided by the user}
}
\value{
One of "normal", "lognormal", "poisson", or "gamma". If there is no
match, an error is thrown.
}
\description{
Use regex to detect a probability distribution name
}
\keyword{internal}
