% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ds_rt.R
\name{get_ds_rt_var}
\alias{get_ds_rt_var}
\title{Calculate the variance for a directly standardized rate}
\usage{
get_ds_rt_var(df)
}
\arguments{
\item{df}{A data frame with the columns \code{popn}, \code{std_popn}, \code{spec_rt}}
}
\value{
A variance as a double
}
\description{
Calculate the variance for a directly standardized rate
}
\examples{
\dontrun{
df <- data.frame(counts, popn, std_popn) \%>\%
  dplyr::mutate(
    spec_rt = get_spec_rt(counts, popn),
    exp_counts = spec_rt * std_popn
  )

get_ds_rt_var(df)
}

}
\keyword{internal}
