% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spec_rt.R
\name{get_single_rt}
\alias{get_single_rt}
\title{Calculate a single rate and evaluate the validity of the output}
\usage{
get_single_rt(x, y)
}
\arguments{
\item{x}{The numerator of the rate}

\item{y}{The denominator of the rate}
}
\value{
A data frame containing the rate and validity of the output.
}
\description{
Calculate a single rate and evaluate the validity of the output
}
\examples{
\dontrun{
a <- 10
b <- 100

get_single_rt(a, b)
}

}
\keyword{internal}
