% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spec_rt.R
\name{get_spec_rt}
\alias{get_spec_rt}
\title{Calculate specific rates}
\usage{
get_spec_rt(
  counts,
  popn,
  scale = NULL,
  power = NULL,
  output_status = FALSE,
  dist = NULL,
  interval = NULL
)
}
\arguments{
\item{counts}{A numeric vector containing the frequency of events.}

\item{popn}{A numeric vector containing the size of each defined population.}

\item{scale, power}{Provide either \code{scale}, a positive integer, to specify the
scale, or \code{power}, a non-negative integer, to specify the power of 10 to use
when expressing the result. If neither are supplied, \code{scale = 1} will be
applied.}

\item{output_status}{If \code{FALSE} (default) the calculated rates are returned
as a numeric vector. If \code{TRUE} a data frame containing both the rates and
the validity of the corresponding elements in the input vectors is returned.}

\item{dist}{A string to indicate the type of probability distribution for the confidence interval to follow.
Can be one of the following: \code{"normal"}, \code{"log normal"}, or \code{"poisson"}. This parameter is case insensitive.}

\item{interval}{A scalar, between 0 and 1, indicating the width of the confidence interval.
For example, for a 95\% confidence interval, use \code{interval = 0.95}.}
}
\value{
If \code{output_status = FALSE} and neither \code{dist} or \code{interval} is supplied, a numeric vector is returned.

Otherwise, a data frame is returned with a column for \code{rate} and up to 4 additional columns:
\itemize{
\item If \code{output_status = TRUE}, then the column \code{status} is included.
\item If both \code{dist} and \code{interval} are provided, then the columns \code{interval}, \code{lower}, and \code{upper} are also included.
}
}
\description{
Calculate specific rates given a vector containing the frequencies of events
and a vector containing the sizes of the defined populations.
}
\details{
This low-level function assumes that counts and sizes of respective
subpopulations are provided in the same order. It can also be used to
calculate a crude rate if \code{counts} and \code{popn} are scalars and represent the
sum of all counts and the size of the entire population respectively.

The \code{status} column of the output can be interpreted as follows:
\itemize{
\item \code{"has_NA"} indicates that the numerator and/or denominator of the fraction
was NA.
\item \code{"denom_0"} indicates that the denominator of the fraction was 0.
\item \code{"success"} indicates that both the numerator and denominator of the
fraction were valid numbers.
}

To construct confidence intervals for the rate estimates, arguments must be supplied to both \code{dist} and \code{interval}.
In general, \emph{the normal distribution should not be used to construct confidence intervals when the standard deviation
is greater than a third of the mean}. The function throws a warning if this is the case.
}
\examples{
\dontrun{
# using vectors
counts <- c(1, 2, 3, 4, 5)
popn <- c(10, 50, 100, 500, 1000)

get_spec_rt(counts, popn)
get_spec_rt(counts, popn, scale = 1000)
get_spec_rt(counts, popn, power = 3)

# get output status
get_spec_rt(counts, popn, output_status = TRUE)

# construct confidence intervals for rate estimates
get_spec_rt(counts, popn, dist = "normal", interval = 0.95)

# using a data frame
df <- data.frame(counts, popn)

get_spec_rt(df$counts, df$popn, scale = 100000)
get_spec_rt(df$counts, df$popn, power = 5)

# using dplyr to create a column
df <- df \%>\%
  dplyr::mutate(spec_rate = get_spec_rt(counts, popn, scale = 1000))
}

}
\references{
\href{http://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_stdrate_details.htm}{The STDRATE Procedure}
}
