% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Study population size}
\format{
A data frame with 190 rows and 4 variables:
\describe{
\item{year}{Year}
\item{age_group}{5-year age group}
\item{sex}{Sex}
\item{popn}{Number of people}
}
}
\usage{
population
}
\description{
A data set containing the number of people in each subgroup of the \emph{simulated}
study population.
}
\keyword{datasets}
