% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std_population}
\alias{std_population}
\title{Standard population size}
\format{
A data frame with 38 rows and 3 variables:
\describe{
\item{age_group}{5-year age group}
\item{sex}{Sex}
\item{std_popn}{Number of people}
}
}
\usage{
std_population
}
\description{
A data set containing the number of people in each subgroup of the \emph{simulated}
standard population.
}
\keyword{datasets}
