% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci_pois}
\alias{get_ci_pois}
\title{Replicate \code{stats::qpois} using \code{stats::qchisq}}
\usage{
get_ci_pois(q_low, q_high, estimate, denominator)
}
\arguments{
\item{q_low}{A scalar, indicating the lower limit}

\item{q_high}{A scalar, indicating the upper limit}

\item{estimate}{A scalar, must be a whole number (e.g. number of events)}

\item{denominator}{A scalar (e.g. population-time)}
}
\value{
A numeric vector containing the lower and upper limits
}
\description{
This function implements the Poisson distribution using the
chi-squared distribution to allow for the continuous extension of integer
variable (\code{estimate}). Read p.9385 of the \href{https://support.sas.com/documentation/onlinedoc/stat/151/stdrate.pdf}{STDRATE Procedure}
for more information.
}
\examples{
\dontrun{
get_ci_pois(q_low = 0.025, q_high = 0.975, estimate = 30, denominator = 1000)
}

}
\keyword{internal}
