% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smr.R
\name{get_smr}
\alias{get_smr}
\title{Perform indirect rate standardization}
\usage{
get_smr(count, popn, std_rt, output_type = "ratio", percent = FALSE)
}
\arguments{
\item{count}{A scalar, the number of observed deaths.}

\item{popn}{A numeric vector containing the size of each defined population.}

\item{std_rt}{A numeric vector containing the specific rate (expressed as
the number of cases per unit of population) for each stratum in the standard
population.}

\item{output_type}{If \code{"ratio"} (default), only the SMR is returned. If
\code{"counts"}, the expected number of deaths for each stratum in the observed
population is returned as a vector.}

\item{percent}{If \code{FALSE} (default), the SMR is returned as is (e.g. 0.5).
If \code{TRUE}, the SMR is expressed as a percentage (e.g. 50).}
}
\value{
Either the SMR as a scalar or an array of the expected counts for
each stratum.
}
\description{
Compute a standardized mortality ratio (SMR) or expected number
of deaths given the number of observed deaths, the size of each stratum in
the observed population, and the specific rate for each stratum in the
standard population.
}
\details{
This low-level function assumes that the size of each stratum in the
observed population and the specific rate for the respective stratum in the
standard population are provided in the same order.
}
\examples{
\dontrun{
count <- 1500
popn <- c(150000, 55000, 12000)
std_rt <- c(0.001, 0.003, 0.05)

get_smr(count, popn, std_rt)
get_smr(count, popn, std_rt, output_type = "ratio")
get_smr(count, popn, std_rt, output_type = "counts")
get_smr(count, popn, std_rt, percent = FALSE)
get_smr(count, popn, std_rt, percent = TRUE)
}

}
